use numerical integration to find number DCS decays + return a double

+++ src/PullStudyHelpers.h
@@ -27,8 +27,10 @@ void plot_parameter_distribution(std::string         title,
  * The ratio of DCS to CF decays is calculated from the ratio of the integrals of their decay rates.
  * CF rate is exponential; DCS is exp * (a + bt + ct^2)
  *
+ * Returns a double so e.g. can be used as the mean of a distribution. Cast to an integer type before using as a count!
+ *
  */
-size_t numDCSDecays(const size_t numCFDecays, const DecayParams_t &phaseSpaceParams, double maxTime);
+double numDCSDecays(const size_t numCFDecays, const DecayParams_t &phaseSpaceParams, double maxTime);
 
 } // namespace PullStudyHelpers
